Some adjustments

+++ src/editor/ui/panels/ResourceSelectionPanel.h
@@ -34,7 +34,7 @@ namespace Atlas::Editor::UI {
             auto buttonName = resourceHandle.IsValid() ? resourceHandle.GetResource()->GetFileName() :
                 "Drop resource here##" + std::to_string(counter++);
 
-            ImGui::PushID(resourceHandle.IsValid() ? resourceHandle.GetID() : counter);
+            ImGui::PushID(resourceHandle.IsValid() ? int32_t(resourceHandle.GetID()) : counter);
 
             popup.SetID(resourceHandle.IsValid() ? resourceHandle.GetID() : counter);
             if (ImGui::Button(buttonName.c_str(), { resourceButtonSize, 0 }))
